updated object searched on prototype chain.
Found a bug in my code that does not logically work the same as the native isPrototypeOf. 

The isPrototypeOf() method allows you to check whether or not an object exists within another object's prototype chain

Comparing the native function, the prototype object passed in the first argument should not be true if the second argument is that same prototype object. (it does not exist within another object's prototype chain)

Now, I immediately search Object.getPrototypeOf on object before starting the while loop. 

This allows for a scenario like isPrototypeOf(james, Object.getPrototypeOf(james2)) to fail the test (since the prototype of james2 is james. and isPrototypeOf(james, james) should fail since the object in comparison is the actual prototype (which does not exist within another objects prototype chain)).